+++ src/RazorLight/Compilation/DefaultAssemblyPathFormatter.cs
@@ -4,6 +4,6 @@ namespace RazorLight.Compilation
 {
 	public class DefaultAssemblyPathFormatter : IAssemblyPathFormatter
 	{
-		public string GetAssemblyPath(Assembly assembly) => assembly.Location;
+		public string GetAssemblyPath(Assembly assembly) => assembly.IsDynamic ? string.Empty : assembly.Location;
 	}
 }