+++ src/RazorLight/Compilation/ICompilationService.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.Emit;
 using RazorLight.Generation;
+using System.Diagnostics.CodeAnalysis;
 using System.Reflection;
 
 namespace RazorLight.Compilation
@@ -12,6 +13,8 @@ namespace RazorLight.Compilation
 		CSharpParseOptions ParseOptions { get; }
 		Assembly OperatingAssembly { get; }
 
+		[RequiresDynamicCode(DeploymentCompatibility.RequiresDynamicCodeMessage, Url = DeploymentCompatibility.DocumentationUrl)]
+		[RequiresUnreferencedCode(DeploymentCompatibility.RequiresUnreferencedCodeMessage, Url = DeploymentCompatibility.DocumentationUrl)]
 		Assembly CompileAndEmit(IGeneratedRazorTemplate razorTemplate);
 	}
 }